Add horizontal line, default colors for appending to buffer
